+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-nacos-discovery/src/main/resources/META-INF/spring.factories
@@ -7,7 +7,7 @@ org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   com.alibaba.cloud.nacos.discovery.reactive.NacosReactiveDiscoveryClientConfiguration,\
   com.alibaba.cloud.nacos.discovery.configclient.NacosConfigServerAutoConfiguration,\
   com.alibaba.cloud.nacos.NacosServiceAutoConfiguration,\
-  com.alibaba.cloud.nacos.intetuntil.InetIPv6Utils
+  com.alibaba.cloud.nacos.intetutil.InetIPv6Utils
 org.springframework.cloud.bootstrap.BootstrapConfiguration=\
   com.alibaba.cloud.nacos.discovery.configclient.NacosDiscoveryClientConfigServiceBootstrapConfiguration
 org.springframework.context.ApplicationListener=\
